useRoomIdの機能をuseRoomに統合しつつ、404 roomの問題を解消した
