+++ src/App.tsx
@@ -1,16 +1,13 @@
 import { Debug } from "./Debug"
 import { Timer } from "./Timer"
 import { useRoom } from "./useRoom"
-import { useRoomId } from "./useRoomId"
 
 export function App() {
-  const roomId = useRoomId()
-  const room = useRoom(roomId)
-  console.log(room)
+  const room = useRoom()
 
   return (
     <>
-      <Timer key={roomId} roomId={roomId} />
+      <Timer key={room.id} roomId={room.id} />
 
       {import.meta.env.DEV && <Debug />}
     </>
