+++ src/useRoom.ts
@@ -1,12 +1,27 @@
-import { doc, getDoc } from "firebase/firestore"
+import {
+  doc,
+  Firestore,
+  getDoc,
+  serverTimestamp,
+  Timestamp,
+  writeBatch,
+} from "firebase/firestore"
 import { collection } from "./collection"
 import { mapGetOrPut } from "./mapGetOrPut"
-import { Room, roomZod } from "./roomZod"
+import { Room, RoomOnFirestore, roomZod } from "./roomZod"
 import { Store } from "./Store"
+import { TimerActionOnFirestore } from "./timerAction"
 import { useFirestore } from "./useFirestore"
+import { setHash, useHash } from "./useHash"
 
-export function useRoom(roomId: string): Room {
+export function useRoom(): Room {
   const db = useFirestore()
+
+  let roomId = useHash().slice("#".length)
+  if (!roomId) {
+    roomId = doc(collection(db, "rooms")).id
+    setHash(roomId)
+  }
 
   const store = getOrPut(
     roomId,
@@ -14,8 +29,16 @@ export function useRoom(roomId: string): Room {
       new Store((onChange) => {
         const abort = new AbortController()
 
-        getDoc(doc(collection(db, "rooms"), roomId)).then((roomDoc) => {
+        const roomRef = doc(collection(db, "rooms"), roomId)
+
+        getDoc(roomRef).then(async (roomDoc) => {
           if (abort.signal.aborted) return
+
+          if (!roomDoc.exists()) {
+            await setupRoom(db, roomId)
+
+            roomDoc = await getDoc(roomRef)
+          }
 
           const room: Room = {
             ...roomZod.parse(roomDoc.data()),
@@ -35,3 +58,29 @@ export function useRoom(roomId: string): Room {
 }
 
 const getOrPut = mapGetOrPut(new Map<Room["id"], Store<Room>>())
+
+async function setupRoom(db: Firestore, newRoomId: string): Promise<void> {
+  const batch = writeBatch(db)
+
+  const rooms = collection(db, "rooms")
+  const newRoom: RoomOnFirestore = {
+    lastEditAt: serverTimestamp() as Timestamp,
+  }
+  batch.set(doc(rooms, newRoomId), {
+    ...newRoom,
+    createdAt: serverTimestamp(),
+  })
+
+  const actions = collection(db, "rooms", newRoomId, "actions")
+  const newActionId = doc(actions).id
+  const newAction: TimerActionOnFirestore = {
+    type: "edit-done",
+    duration: 3 * 60_000,
+  }
+  batch.set(doc(actions, newActionId), {
+    ...newAction,
+    createdAt: serverTimestamp(),
+  })
+
+  await batch.commit()
+}
